requirements.txt
flask==2.3.3
Flask-SQLAlchemy
flask-login
fuzzywuzzy
werkzeug==2.3.7